% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number.R
\name{get_number}
\alias{get_number}
\alias{get_number_vec}
\alias{get_numbers}
\alias{get_numbers_vec}
\title{Conversion functions}
\usage{
get_number()

get_number_vec()

get_numbers()

get_numbers_vec()
}
\value{
Modifies current document
}
\description{
Conversion functions
}
